requirements.txt
requests==2.31.0
python-dotenv==1.0.0
SQLAlchemy==2.0.25
plotly==5.18.0
dash==2.14.2
pandas==2.1.4